+++ lib/timely/rails/season.rb
@@ -1,16 +1,10 @@
 module Timely
   class Season < ActiveRecord::Base
-
-    if ::ActiveRecord::VERSION::MAJOR >= 4
-      has_many :date_groups, lambda { order(:start_date) }, :dependent => :destroy, :class_name => 'Timely::DateGroup'
-    else
-      has_many :date_groups, :order => :start_date, :dependent => :destroy, :class_name => 'Timely::DateGroup'
-    end
-
+    has_many :date_groups, -> { order(:start_date) }, dependent: :destroy, class_name: 'Timely::DateGroup', inverse_of: :season
 
     accepts_nested_attributes_for :date_groups,
-      :reject_if => proc {|attributes| attributes['start_date'].blank?},
-      :allow_destroy => true
+      reject_if: proc { |attributes| attributes['start_date'].blank? },
+      allow_destroy: true
 
     validate :validate_dates_specified
 
@@ -21,14 +15,6 @@ module Timely
 
     def includes_date?(date)
       date_groups.any?{|dg| dg.includes_date?(date)}
-    end
-
-    def has_gaps?
-      last_date = nil
-      date_groups.each do |dg|
-        return true if last_date && dg.start_date != last_date + 1
-      end
-      false
     end
 
     def dates
@@ -67,15 +53,6 @@ module Timely
       cloned
     end
 
-    def self.build_season_for(dates=[])
-      season = Season.new
-      date_groups = dates.map do |date|
-        DateGroup.new(:start_date => date, :end_date => date)
-      end
-      season.date_groups = date_groups
-      season
-    end
-
     def to_s
       name.presence || Timely::DateRange.to_s(boundary_start, boundary_end)
     end
@@ -83,7 +60,9 @@ module Timely
     alias_method :audit_name, :to_s
 
     def string_of_date_groups
-      date_groups.map{|dg| "#{dg.start_date.to_s(:short)} - #{dg.end_date.to_s(:short)}"}.to_sentence
+      date_groups.map{ |dg|
+        "#{dg.start_date.to_s(:short)} - #{dg.end_date.to_s(:short)}"
+      }.to_sentence
     end
   end
 end