+++ lib/timely/rails/date_group.rb
@@ -2,10 +2,11 @@
 
 module Timely
   class DateGroup < ActiveRecord::Base
-    belongs_to :season, class_name: 'Timely::Season', optional: true
+    belongs_to :season, class_name: 'Timely::Season', optional: true, inverse_of: :date_groups
 
     weekdays_field :weekdays
 
+    validates :weekdays_bit_array, presence: true
     validates_presence_of :start_date, :end_date
     validate :validate_date_range!
 
@@ -23,9 +24,7 @@ module Timely
     }
 
     scope :for_any_weekdays, lambda { |weekdays_int|
-      weekdays_int = weekdays_int.to_i
-      where((arel_table[:weekdays_bit_array] & weekdays_int).not_eq(0))
-        .or(where(weekdays_bit_array: nil))
+      where((arel_table[:weekdays_bit_array] & weekdays_int.to_i).not_eq(0))
     }
 
     scope :applying_for_duration, lambda { |date_range|