+++ spec/date_group_spec.rb
@@ -9,6 +9,11 @@ RSpec.describe Timely::DateGroup do
     )
   end
 
+  it 'is not valid to set nil weekdays' do
+    @date_group.weekdays_bit_array = nil
+    expect(@date_group).not_to be_valid
+  end
+
   it 'should detect overlaps' do
     expect(@date_group.applicable_for_duration?(Timely::DateRange.new('2000-01-01'.to_date, '2000-01-01'.to_date))).to be true
     expect(@date_group.applicable_for_duration?(Timely::DateRange.new('2000-01-01'.to_date, '2000-01-06'.to_date))).to be true
@@ -28,23 +33,19 @@ RSpec.describe Timely::DateGroup do
   end
 end
 
+RSpec.describe Timely::DateGroup, 'without weekdays' do
+  subject { Timely::DateGroup.new(start_date: Date.current, end_date: Date.current) }
+  it { is_expected.to be_valid }
+end
+
 RSpec.describe 'Timely::DateGroup.for_any_weekdays' do
   let(:date_range) { ('2019-10-17'.to_date)..('2019-10-18'.to_date) }
   let(:weekdays_int) { Timely::WeekDays.from_range(date_range).weekdays_int }
 
-  let(:special_group) do
-    Timely::DateGroup.create(start_date: date_range.first, end_date: date_range.last).tap do |date_group|
-      date_group.update_column(:weekdays_bit_array, nil)
-    end.reload
-  end
-
-  let!(:date_groups) do
-    [
-      Timely::DateGroup.create(start_date: date_range.first, end_date: date_range.last, weekdays: { thu: true }),
-      Timely::DateGroup.create(start_date: date_range.first, end_date: date_range.last, weekdays: { mon: true }),
-      special_group
-    ]
-  end
+  let!(:date_groups) { [
+    Timely::DateGroup.create(start_date: date_range.first, end_date: date_range.last, weekdays: { thu: true }),
+    Timely::DateGroup.create(start_date: date_range.first, end_date: date_range.last, weekdays: { mon: true }),
+  ] }
 
   RSpec.shared_examples 'finds expected date groups' do
     it 'finds expected date groups' do
@@ -63,14 +64,14 @@ RSpec.describe 'Timely::DateGroup.for_any_weekdays' do
 
   context '#within_range' do
     let(:scoped_result) { Timely::DateGroup.within_range(date_range).to_a }
-    let(:expected_groups) { [date_groups[0], date_groups[1], date_groups[2]] }
+    let(:expected_groups) { [date_groups[0], date_groups[1]] }
     let(:absent_groups) { [] }
     let(:includes_date_groups) { [] }
     it_behaves_like 'finds expected date groups'
   end
 
   let(:includes_date_groups) { expected_groups }
-  let(:expected_groups) { [date_groups[0], date_groups[2]] }
+  let(:expected_groups) { [date_groups[0]] }
   let(:absent_groups) { [date_groups[1]] }
 
   context '#for_any_weekdays' do